+++ src/js/helpers/input-trailing-digits-count.js
@@ -2,12 +2,12 @@ import config from '@/config'
 
 export function inputStepByDigitsCount (count) {
   const defaultStr = '0'
-  if (+count > 0 && +count <= config.DECIMAL_POINTS) {
+  if (+count > 0 && +count <= config.MAX_SYSTEM_DECIMAL_POINTS) {
     return `${defaultStr}.${'1'.padStart(+count, '0')}`
   } else if (+count === 0) {
     return defaultStr
-  } else if (+count > config.DECIMAL_POINTS || +count < 0) {
-    throw new Error(`Trailing digits count for this asset should be non-negative and less than ${config.DECIMAL_POINTS}`)
+  } else if (+count > config.MAX_SYSTEM_DECIMAL_POINTS || +count < 0) {
+    throw new Error(`Trailing digits count for this asset should be non-negative and less than ${config.MAX_SYSTEM_DECIMAL_POINTS}`)
   }
 }